fix compiler warnings and EToday bounds check
 fix off-by-one bounds check in record loop: recsize=16 records need
  only 16 bytes headroom, not 20; previous check cut loop short,
  causing SPOT_ETODAY to never be read
- fix is_NaN(int16_t): cast to uint16_t before comparing NaN_S16
  (0x8000 out of signed int16_t range → always-false comparison)
- remove const qualifier from is_NaN return types (-Wignored-qualifiers)
- fix printf format specifiers: %d→%ld for int32_t, %u/%X→%lu/%lX
  for uint32_t on platforms where uint32_t=long unsigned int

+++ esphome/components/smabluetooth_solar/smabluetooth_solar.cpp
@@ -293,7 +293,7 @@ void SmaBluetoothSolar::updateSensor(binary_sensor::BinarySensor *sensor, const 
 void SmaBluetoothSolar::dump_config() {
     ESP_LOGCONFIG(TAG, "SMABluetooth Solar:");
     ESP_LOGCONFIG(TAG, "  MAC: %s", sma_inverter_bluetooth_mac_.c_str());
-    ESP_LOGCONFIG(TAG, "  Delay between queries: %u ms", sma_inverter_delay_values_);
+    ESP_LOGCONFIG(TAG, "  Delay between queries: %lu ms", (unsigned long)sma_inverter_delay_values_);
 }
 
 const char *SmaBluetoothSolar::lookup_code(uint16_t code) {
